requirements.txt
# Główne zależności aplikacji
streamlit>=1.28.0          # Framework do tworzenia interfejsu webowego
openai>=1.0.0              # Komunikacja z API OpenAI (GPT + TTS)
python-docx>=1.0.0         # Tworzenie i odczyt plików Word (.docx)
python-dotenv>=1.0.0       # Ładowanie zmiennych środowiskowych z pliku .env
requests>=2.31.0           # Komunikacja HTTP z Vercel Blob API
pydub>=0.25.1              # Manipulacja plikami audio (łączenie, przerwy)
